requirements.txt
wheel
dists-pytorch==0.1
diffusers==0.25.1
transformers==4.37.2
tqdm
pytorch_lightning==2.2.5
black
ruff
hydra-core
jaxtyping
beartype
wandb
einops
colorama
scikit-image
colorspacious
matplotlib
moviepy
imageio
git+https://github.com/Chrixtar/latent-gaussian-rasterization
timm
dacite
lpips
e3nn
plyfile
tabulate
svg.py
opencv-python==4.6.0.66
sk-video
natsort==8.4.0
numpy==1.26.0